fix: actionJs button main export
