+++ src/prefabs/actionJsButton.tsx
@@ -1,0 +1,12 @@
+import { Icon, prefab } from '@betty-blocks/component-sdk';
+import { ActionJSButton } from './structures/ActionJSButton';
+
+const attributes = {
+  category: 'FORMV2',
+  icon: Icon.ButtonIcon,
+  keywords: ['Form', 'input'],
+};
+
+export default prefab('ActionJs Button', attributes, undefined, [
+  ActionJSButton({}),
+]);
