+++ src/prefabs/structures/ActionJSButton/options/index.ts
@@ -23,6 +23,9 @@ export const options = {
   actionId: option('ACTION_JS', {
     label: 'Action',
     value: '',
+    configuration: {
+      disabled: true,
+    },
   }),
 
   property: property('Pass value to action'),
